feat: 휴가 승인시 잔여 휴가에서 차감
annualLeave int->double

+++ src/main/java/com/example/exodia/user/dto/UserProfileDto.java
@@ -19,7 +19,7 @@ public class UserProfileDto {
     private LocalDate joinDate;
     private String name;
     private String profileImage;
-    private int annualLeave;
+    private double annualLeave;
 
     public static UserProfileDto fromProfileEntity(User user) {
         return new UserProfileDto(
